Move shared code into AbstractConfigManager
