+++ src/YamlConfigManager.php
@@ -9,25 +9,8 @@ use RuntimeException;
  * Class YamlConfigManager
  * @package clagiordano\weblibs\configmanager
  */
-class YamlConfigManager implements IConfigurable
+class YamlConfigManager extends AbstractConfigManager
 {
-    /** @var array $configData */
-    private $configData = null;
-    /** @var string $configFilePath */
-    private $configFilePath = null;
-
-    /**
-     * Create config object, optionally automatic load config
-     * from argument $configFilePath
-     *
-     * @param string $configFilePath
-     * @return IConfigurable
-     */
-    public function __construct($configFilePath = null)
-    {
-        return $this->loadConfig($configFilePath);
-    }
-
     /**
      * Load config data from file and store it into internal property
      *
@@ -74,75 +57,6 @@ class YamlConfigManager implements IConfigurable
         if ($autoReloadConfig) {
             $this->loadConfig($configFilePath);
         }
-
-        return $this;
-    }
-
-    /**
-     * Get value pointer from config for get/set value
-     *
-     * @param string $configPath
-     *
-     * @return mixed
-     */
-    private function & getValuePointer($configPath)
-    {
-        $configData =& $this->configData;
-        $parts = explode('.', $configPath);
-        $length = count($parts);
-
-        for ($i = 0; $i < $length; $i++) {
-            if (!isset($configData[ $parts[ $i ] ])) {
-                $configData[ $parts[ $i ] ] = ($i === $length) ? [] : null;
-            }
-
-            $configData = &$configData[ $parts[ $i ] ];
-        }
-
-        return $configData;
-    }
-
-    /**
-     * Get value from config data throught keyValue path
-     *
-     * @param string $configPath
-     * @param mixed $defaultValue
-     *
-     * @return mixed
-     */
-    public function getValue($configPath, $defaultValue = null)
-    {
-        $stored = $this->getValuePointer($configPath);
-
-        return (is_null($stored)
-            ? $defaultValue
-            : $stored);
-    }
-
-    /**
-     * Check if exist required config for keyValue
-     *
-     * @param string $keyValue
-     *
-     * @return mixed
-     */
-    public function existValue($keyValue)
-    {
-        return !is_null($this->getValue($keyValue));
-    }
-
-    /**
-     * Set value in config path
-     *
-     * @param string $configPath
-     * @param mixed $newValue
-     *
-     * @return IConfigurable
-     */
-    public function setValue($configPath, $newValue)
-    {
-        $configData = &$this->getValuePointer($configPath);
-        $configData = $newValue;
 
         return $this;
     }
